+++ apps/web/client/eslint.config.js
@@ -30,8 +30,8 @@ export default tseslint.config(
       ],
       "@typescript-eslint/require-await": "off",
       "@typescript-eslint/no-misused-promises": [
-        "error",
-        { checksVoidReturn: { attributes: false } },
+        "warn",
+        { checksVoidReturn: { attributes: false, functions: false, methods: false } },
       ],
     },
   },
