+++ packages/fonts/src/helper.ts
@@ -1,105 +1,5 @@
-import type { Font } from "@onlook/models";
-import { camelCase } from "lodash";
 import * as t from "@babel/types";
-
-export function extractFontConfig(
-  fontId: string,
-  fontType: string,
-  configArg: t.ObjectExpression,
-): Font {
-  const fontConfig: Record<string, any> = {
-    id: fontId,
-    family: fontType === "localFont" ? fontId : fontType.replace(/_/g, " "),
-    type: fontType === "localFont" ? "local" : "google",
-    subsets: [],
-    weight: [],
-    styles: [],
-    variable: "",
-  };
-
-  configArg.properties.forEach((prop) => {
-    if (!t.isObjectProperty(prop) || !t.isIdentifier(prop.key)) {
-      return;
-    }
-
-    const propName = prop.key.name;
-
-    if (propName === "variable" && t.isStringLiteral(prop.value)) {
-      fontConfig.variable = prop.value.value;
-    }
-
-    if (propName === "subsets" && t.isArrayExpression(prop.value)) {
-      fontConfig.subsets = prop.value.elements
-        .filter((element): element is t.StringLiteral =>
-          t.isStringLiteral(element),
-        )
-        .map((element) => element.value);
-    }
-
-    if (
-      (propName === "weight" || propName === "weights") &&
-      t.isArrayExpression(prop.value)
-    ) {
-      fontConfig.weight = prop.value.elements
-        .map((element) => {
-          if (t.isStringLiteral(element)) {
-            return element.value;
-          } else if (t.isNumericLiteral(element)) {
-            return element.value.toString();
-          }
-          return null;
-        })
-        .filter(
-          (weight): weight is string =>
-            weight !== null && !isNaN(Number(weight)),
-        );
-    }
-
-    if (
-      (propName === "style" || propName === "styles") &&
-      t.isArrayExpression(prop.value)
-    ) {
-      fontConfig.styles = prop.value.elements
-        .filter((element): element is t.StringLiteral =>
-          t.isStringLiteral(element),
-        )
-        .map((element) => element.value);
-    }
-
-    // Handle local font src property
-    if (
-      propName === "src" &&
-      t.isArrayExpression(prop.value) &&
-      fontType === "localFont"
-    ) {
-      const srcConfigs = prop.value.elements
-        .filter((element): element is t.ObjectExpression =>
-          t.isObjectExpression(element),
-        )
-        .map((element) => {
-          const srcConfig: Record<string, string> = {};
-          element.properties.forEach((srcProp) => {
-            if (t.isObjectProperty(srcProp) && t.isIdentifier(srcProp.key)) {
-              const srcPropName = srcProp.key.name;
-              if (t.isStringLiteral(srcProp.value)) {
-                srcConfig[srcPropName] = srcProp.value.value;
-              }
-            }
-          });
-          return srcConfig;
-        });
-
-      fontConfig.weight = [
-        ...new Set(srcConfigs.map((config) => config.weight)),
-      ];
-      fontConfig.styles = [
-        ...new Set(srcConfigs.map((config) => config.style)),
-      ];
-    }
-  });
-
-  return fontConfig as Font;
-}
+import { camelCase } from "lodash";
 
 /**
  * Helper function to find font class in a string of class names
@@ -268,4 +168,3 @@ export function removeFontsFromClassName(
 
   return modified;
 }
-
