+++ packages/fonts/src/index.ts
@@ -1,3 +1,5 @@
 export * from './default';
 export * from './family';
 export * from './variants';
+export * from './helper';
+export * from './utils';
